refactor: domain 패키지 생성
- 기존에는 UI 관련한 객체들을 모아놓는 view패키지가 존재했다.
이번에는 핵심 비즈니스 로직을 가지는 도메인 객체들을 모아놓는
domain 패키지를 생성한다.
